Move intrange to its own module
types is pretty generic as a module name, this should be split into
seperate modules.
